Fixed total cost on date
- Did a bit of refactoring
- Tidied imports
