+++ buffetiser/backend/core/admin.py
@@ -1,6 +1,7 @@
+from django.contrib import admin
+
 from core.models import (DailyChange, DividendPayment, DividendReinvestment,
                          History, Investment, Purchase, Sale)
-from django.contrib import admin
 
 # Register your models here.
 admin.site.register(Investment)
