+++ buffetiser/backend/core/management/commands/bulk_import_dividends.py
@@ -11,8 +11,9 @@ e.g. 1/11/2022, CB, 11 VAS @ 85.4600 AUD, Vanguard Australian Shares Index ETF, 
 import csv
 from datetime import datetime
 
-from core.models import Investment, Purchase, Sale
 from django.core.management.base import BaseCommand
+
+from core.models import Investment, Purchase, Sale
 
 
 class Command(BaseCommand):
