+++ buffetiser/backend/core/models.py
@@ -1,8 +1,9 @@
 from typing import Union
 
+from django.db.models import (CASCADE, BooleanField, CharField, DateField,
+                              FloatField, ForeignKey, IntegerField, Model)
+
 from core.config import Constants
-from django.db.models import (CASCADE, CharField, DateField, FloatField,
-                              ForeignKey, IntegerField, Model, BooleanField)
 
 
 def date_to_string(date):
