+++ buffetiser/backend/core/serializers.py
@@ -1,6 +1,7 @@
+from rest_framework import serializers
+
 from core.models import (DailyChange, DividendPayment, DividendReinvestment,
                          History, Investment, Purchase, Sale)
-from rest_framework import serializers
 
 
 class InvestmentSerializer(serializers.ModelSerializer):
