+++ buffetiser/backend/core/services/investment_helpers.py
@@ -3,7 +3,8 @@ import datetime
 
 import aiohttp
 from asgiref.sync import sync_to_async
-from core.models import Investment, Purchase, Sale
+
+from core.models import Investment, Sale
 
 
 def date_to_string(date):
@@ -27,56 +28,11 @@ def fe_string_to_date(date_string):
     return datetime.datetime.strptime(date_string, "%Y-%m-%d")
 
 
-def get_purchase_history(investment):
+def nearest(items, pivot):
     """
-    All the purchases that have been made of this Investment.
+    Find the nearest item (date) in a list.
     """
-    investment_purchase_history = Purchase.objects.filter(investment=investment).all()
-    purchases = {}
-    for purchase in investment_purchase_history:
-        purchases.setdefault(purchase.date, []).append(
-            {
-                "units": purchase.units,
-                "price_per_unit": purchase.price_per_unit,
-                "total_cost": purchase.units * purchase.price_per_unit,
-            }
-        )
-    return purchases
-
-
-def get_sale_history(investment):
-    """
-    All the sales that have been made of this Investment.
-    """
-    investment_sale_history = Sale.objects.filter(investment=investment).all()
-    sales = {}
-    for sale in investment_sale_history:
-        sales.setdefault(sale.date, []).append(
-            {
-                "units": sale.units,
-                "price_per_unit": sale.price_per_unit,
-                "total_cost": sale.units * sale.price_per_unit,
-            }
-        )
-    return sales
-
-
-def get_units_held_at_date(investment, cut_off_date):
-    """
-    This gets all the units held for an Investment up to (and including) a certain date.
-    """
-    purchases = get_purchase_history(investment)
-    sales = get_sale_history(investment)
-
-    units_held_at_date = 0
-    for purchase_date in list(purchases.keys()):
-        if datetime.datetime.strptime(purchase_date, "%d/%m/%Y") <= cut_off_date:
-            units_held_at_date += purchases[purchase_date][0][0]
-    for sales_date in list(sales.keys()):
-        if datetime.datetime.strptime(sales_date, "%d/%m/%Y") <= cut_off_date:
-            units_held_at_date -= purchases[purchase_date][0][0]
-
-    return int(units_held_at_date)
+    return min(items, key=lambda x: abs(x - pivot.date()))
 
 
 # ***********************************************
